refactor: change character type
