+++ app/src/main/res/layout/activity_character_detail_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.CharacterDetailScreenActivity">
+    tools:context=".ui.detail.CharacterDetailScreenActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
