+++ app/src/main/res/layout/activity_character_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.CharacterListActivity">
+    tools:context=".ui.list.CharacterListActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/character_list_toolbar_layout"
